Clean up nav related components and fix footer after botched merge
